Replace spectron with playwright
