+++ .pre-commit-config.yaml
@@ -5,7 +5,6 @@ repos:
           - id: trailing-whitespace
           # -   id: end-of-file-fixer
           - id: check-yaml
-          - id: check-added-large-files
           - id: check-json
     - repo: https://github.com/python-jsonschema/check-jsonschema
       rev: '0.28.0'
